==> 01 Relevant Data Elements.ipynb <==
# Relevant Data Elements

import csv
import pandas

### MUHC Dataset

# Import original dataset
muhc = pandas.read_csv('00 MUHC Original.csv', delimiter=',', encoding='latin-1')
muhc[:1]

# Create a temporary DataFrame of only the relevant columns.
temp = muhc[['event_type']].copy()
temp['ID'] = muhc['incident_id']
temp['Incident Description'] = muhc['descriptor'] + ' . ' + muhc['incident_description'] # id for incident description
temp['Process Step'] = muhc['process_step_occurred']
temp['Problem Type'] = muhc['problem_type']
temp['Contributing Factors'] = muhc['contributing_factors']
temp['Overall Severity'] = muhc['acute_medical_harm']
muhc = temp[['ID', 'Process Step', 'Problem Type', 'Contributing Factors', 'Overall Severity', 'Incident Description']]
muhc[:1]

# Save the new dataset as '2 MUHC subset' because
    # this is the second version of the MUHC dataset
    # the columns of this version are a subset of the original version
muhc.to_csv('01 MUHC Relevant.csv', encoding='latin-1', index=False, quoting=csv.QUOTE_NONNUMERIC)

### CIHI Dataset

cihi = pandas.read_csv('00 CIHI Original.csv', delimiter=',', encoding='latin-1')
cihi[:1]

temp = cihi[['Frequency']].copy()
temp['ID'] = cihi['RT NSIR Case ID']
temp['Incident Description'] = cihi['Incident Description']
temp['Process Step'] = cihi['Process Step Where Incident Occurred']
temp['Problem Type'] = cihi['Problem Type - Primary']
temp['Contributing Factors'] = cihi['Contributing Factor List']
temp['Overall Severity'] = cihi['Overall Severity']
cihi = temp[['ID', 'Process Step', 'Problem Type', 'Contributing Factors', 'Overall Severity', 'Incident Description']]
cihi[:1]

cihi.to_csv('01 CIHI Relevant.csv', encoding='latin-1', index=False, quoting=csv.QUOTE_NONNUMERIC)
