==> 02 NSIR-RT Mapping.ipynb <==
# NSIR-RT Mapping
The MUHC data uses the 2016 version of the NSIR-RT, while the CIHI data uses the 2017 version. Here, we map the MUHC data to the CIHI data.

import csv
import pandas

muhc = pandas.read_csv('01 MUHC Relevant.csv', delimiter=',', encoding='latin-1').fillna('')
cihi = pandas.read_csv('01 CIHI Relevant.csv', delimiter=',', encoding='latin-1').fillna('')

### Process Step

set(muhc['Process Step'])

set(cihi['Process Step'])

ps_mapping = {}

# From MUHC to CIHI
ps_mapping['Imaging for radiotherapy planning'] = 'Imaging for treatment planning'
ps_mapping['On-treatment quality management'] = 'On-treatment quality assurance'
ps_mapping['Patient assessment or consultation'] = 'Patient medical consultation and physician assessment'
ps_mapping['Pre-treatment review and verification'] = 'Pre-treatment quality assurance'
ps_mapping['Treatment planning'] = 'Contouring and planning'

# From CIHI to CIHI
ps_mapping['Not Applicable'] = ''
ps_mapping['Patient assessment/consultation (Retired Value)'] = 'Patient medical consultation and physician assessment'
ps_mapping['Post-treatment completion '] = 'Post-treatment completion'
ps_mapping['Other'] = ''

def map_value(mapping, old_value):
    try:
        return mapping[old_value]
    except Exception:
        return old_value

muhc['Process Step'] = muhc['Process Step'].apply(lambda x:map_value(ps_mapping, x))
cihi['Process Step'] = cihi['Process Step'].apply(lambda x:map_value(ps_mapping, x))

### Problem Type

set(muhc['Problem Type'])

set(cihi['Problem Type'])

pt_mapping = {}

# From MUHC to CIHI
pt_mapping['Calculation error'] = 'Wrong prescription dose-fractionation or calculation error'
pt_mapping['Combined modality treatment scheduling error'] = 'Radiation therapy scheduling error'
pt_mapping['Fall or other accident'] = 'Fall or other patient injury or medical condition'
pt_mapping['Hardware/Software'] = 'Systematic hardware/software (including dose-volume) error'
pt_mapping['Interventional procedure error'] = 'Interventional procedure error (Retired value)'
pt_mapping['Radiation treatment scheduling error'] = 'Radiation therapy scheduling error'
pt_mapping['Wrong anatomical site'] = 'Wrong anatomical site (excluding laterality)'
pt_mapping['Wrong patient position'] = 'Wrong patient position, setup point, or shift'
pt_mapping['Wrong prescription dose'] = 'Wrong plan dose (Retired value)'
pt_mapping['Wrong shift from setup point'] = 'Wrong patient position, setup point, or shift'
pt_mapping['Wrong target or OAR contours, or wrong planning margins'] = 'Wrong target or OAR contours or wrong planning (Retired Value)'
pt_mapping['Wrong treatment accessories'] = 'Wrong, missing, mislabeled, or damaged treatment accessories'
pt_mapping['Other'] = ''
pt_mapping[''] = ''

muhc['Problem Type'] = muhc['Problem Type'].apply(lambda x:map_value(pt_mapping, x))
cihi['Problem Type'] = cihi['Problem Type'].apply(lambda x:map_value(pt_mapping, x))

### Contributing Factors

muhc_cfs = []
for cfs in set(muhc['Contributing Factors']):
    for cf in cfs.split('&'):
        muhc_cfs.append(cf.strip())

set(muhc_cfs)

cihi_cfs = []
for cfs in set(cihi['Contributing Factors']):
    temp = cfs.replace("Equipment software or hardware design, including 'human factors' design, inadequate", "Equipment software or hardware design including 'human factors' design inadequate")
    temp = temp.replace("Equipment software or hardware commissioning, calibration or acceptance testing inadequate", "Equipment software or hardware commissioning calibration or acceptance testing inadequate")
    temp = temp.replace("Patient or family member medical condition, preference or behaviour ", "Patient or family member medical condition preference or behaviour")
    for cf in temp.split(','):
        cihi_cfs.append(cf.strip())

set(cihi_cfs)

cf_mapping = {}

# From MUHC to CIHI
cf_mapping['']= ''
cf_mapping['Capital resources inadequate']= 'Organizational and/or workspace resources inadequate (excluding human resources)'
cf_mapping['Commissioning or acceptance testing inadequate']= 'Equipment software or hardware commissioning, calibration or acceptance testing inadequate'
cf_mapping['Communication inappropriate or misdirected']= 'Communication or documentation inadequate (patient specific)'
cf_mapping['Communication not timely']= 'Communication or documentation inadequate (patient specific)'
cf_mapping['Conflicted']= ''
cf_mapping['Documentation poor, incomplete, unclear or missing']= 'Communication or documentation inadequate (patient specific)'
cf_mapping['Education or training inadequate']= 'Staff education or training inadequate'
cf_mapping['Expectation bias']= 'Expectation bias involving staff'
cf_mapping['External factors beyond programmatic control']= 'External factors beyond programmatic control'
cf_mapping['Failure to recognize a hazard']= 'Failure to identify potential risks'
cf_mapping['Failure to select the correct rule']= 'Policies and/or procedures not followed'
cf_mapping['Human behavior involving staff']= 'Staff behaviour'
cf_mapping['Human resources inadequate']= 'Human resources inadequate'
cf_mapping['Inadequate needs or risk assessment']= 'Failure to identify potential risks'
cf_mapping['Loss of attention']= 'Distraction or diversions involving staff'
cf_mapping['Materials, tools or equipment inadequate or insufficient']= 'Organizational and/or workspace resources inadequate (excluding human resources)'
cf_mapping['Medical record poor, incomplete, unclear or missing']= 'Communication or documentation inadequate (patient specific)'
cf_mapping['Non-existent']= ''
cf_mapping['Not followed']= 'Policies and/or procedures not followed'
cf_mapping['Other']= ''
cf_mapping['Patient related circumstances']= 'Patient or family member medical condition preference or behaviour'
cf_mapping['Physical environment inadequate']= 'Organizational and/or workspace resources inadequate (excluding human resources)'
cf_mapping['Plan forgotten in progress']= 'Policies and/or procedures not followed'
cf_mapping['Software operational error']= 'Policies and/or procedures not followed'

# From CIHI to CIHI
cf_mapping['Equipment software or hardware design including human factors design inadequate'] = 'Equipment software or hardware design, including \'human factors\' design, inadequate'
cf_mapping['Equipment software or hardware commissioning calibration or acceptance testing inadequate'] = 'Equipment software or hardware commissioning, calibration or acceptance testing inadequate'
cf_mapping['Patient or family member medical condition preference or behaviour'] = 'Patient or family member medical condition, preference or behaviour'

def map_muhc_cfs(cfs):
    mapped_values = []
    for cf in cfs.split('&'):
        try:
            mapped_values.append(cf_mapping[cf.strip()])
        except Exception:
            mapped_values.append(cf.strip().replace(',', '').replace('\'', ''))
    return '|'.join(mapped_values)

def map_cihi_cfs(cfs):
    
    cf_list = []
    temp = cfs.replace('Equipment software or hardware design, including \'human factors\' design, inadequate', 'Equipment software or hardware design including human factors design inadequate')
    temp = temp.replace('Equipment software or hardware commissioning, calibration or acceptance testing inadequate', 'Equipment software or hardware commissioning calibration or acceptance testing inadequate')
    temp = temp.replace('Patient or family member medical condition, preference or behaviour', 'Patient or family member medical condition preference or behaviour')
    for cf in temp.split(','):
        cf_list.append(cf.strip())
        
    mapped_cfs = []
    for cf in cf_list:
        try:
            mapped_cfs.append(cf_mapping[cf.strip()])
        except Exception:
            mapped_cfs.append(cf.strip())
    return '|'.join(mapped_cfs)

muhc['Contributing Factors'] = muhc['Contributing Factors'].apply(lambda x:map_muhc_cfs(x))
cihi['Contributing Factors'] = cihi['Contributing Factors'].apply(lambda x:map_cihi_cfs(x))

### Overall Severity

set(muhc['Overall Severity'])

set(cihi['Overall Severity'])

os_mapping = {}

# From CIHI to CIHI
os_mapping['Not Applicable'] = ''

muhc['Overall Severity'] = muhc['Overall Severity'].apply(lambda x:map_value(os_mapping, x))
cihi['Overall Severity'] = cihi['Overall Severity'].apply(lambda x:map_value(os_mapping, x))

muhc.to_csv('02 MUHC Mapped.csv', encoding='utf-8', index=False, quoting=csv.QUOTE_NONNUMERIC)
cihi.to_csv('02 CIHI Mapped.csv', encoding='utf-8', index=False, quoting=csv.QUOTE_NONNUMERIC)
