==> 03 Combining Datasets.ipynb <==
# Combining Datasets

import csv
import pandas
pandas.set_option('display.max_colwidth', 0)

muhc = pandas.read_csv('02 MUHC Mapped.csv', delimiter=',', encoding='latin-1').fillna('')
muhc[:1]

cihi = pandas.read_csv('02 CIHI Mapped.csv', delimiter=',', encoding='latin-1').fillna('')
cihi[:1]

combined = pandas.concat([muhc, cihi])
combined[:1]

combined.to_csv('03 Combined.csv', encoding='utf-8', index=False, quoting=csv.QUOTE_NONNUMERIC)
