==> 04 Translating French Incidents.ipynb <==
# Translating French Incidents

import csv
import pandas
import time
from spacy.lang.en.stop_words import STOP_WORDS
from googletrans import Translator
translator = Translator()
pandas.set_option('display.max_colwidth', 0)

combined = pandas.read_csv('03 Combined.csv', delimiter=',', encoding='latin-1').fillna('')
combined['Incident Description'] = combined['Incident Description'].apply(lambda x:x.replace('\r', '.').replace('\n', '.'))
combined[:1]

translations = pandas.read_csv('JK_translations.csv', delimiter=',', encoding='latin-1').fillna('')
translations

# Compile list of stop words with more than 2 characters.
stop_words = [word for word in STOP_WORDS if len(word)>2]

# Returns True or False
def has_english_stop_words(string):
    return len([word for word in set(string.lower().split()) if word in stop_words])>0

# Detect wheterh incident descriptions are in English or French
combined['Language'] = ''
for i in range(len(combined)):
    incident_description = combined.at[i, 'Incident Description']
    if has_english_stop_words(incident_description):
        combined.at[i, 'Language'] = 'En'
    else:
        if translator.detect(incident_description).lang == 'en':
            combined.at[i, 'Language'] = 'En'
        else:
            combined.at[i, 'Language'] = 'Fr'

def translate(string):
    while True:
        try:
            en = translator.translate(string).text
            print(en)
            return en
        except Exception:
            time.sleep(1)

# If the original incident was in English, then 
combined['Translated'] = ''
for i in range(len(combined)):
    incident_description = combined.at[i, 'Incident Description']
    language = combined.at[i, 'Language']
    if language == 'En':
        combined.at[i, 'Translated'] = incident_description
    else:
        translation = translate(incident_description)
        combined.at[i, 'Translated'] = translation

combined[:1]

combined['Language']

combined.to_csv('04 Combined Translated.csv', encoding='utf-8', index=False, quoting=csv.QUOTE_NONNUMERIC)
