==> 07 Data Vectorization.ipynb <==
# Data Vectorization

import csv
import pandas
from sklearn.feature_extraction.text import TfidfVectorizer

def return_tfidf_matrix(column, mx, mn, ng):
    tfidf = TfidfVectorizer(max_df=mx, min_df=mn, ngram_range=ng)
    X = tfidf.fit_transform(column)
    X = pandas.DataFrame.from_records(X.todense().tolist(), columns=tfidf.get_feature_names())
    return X

def return_one_hot(column, penalty):
    option_set = list(set(list(column)))
    return pandas.DataFrame([[1 if value == option else penalty for option in option_set] for value in column], columns = option_set)

### NM PS

dropna_ps = pandas.read_csv('06 NM PS.csv', delimiter=',', encoding='latin-1').fillna('')
dropna_ps[:1]

dropna_ps_tfidf = return_tfidf_matrix(dropna_ps['Bony Cleaned'], 1000, 5, (1,3))
dropna_ps_tfidf.to_csv('07 NM PS TFIDF.csv', encoding='utf-8', index=False, quoting=csv.QUOTE_NONNUMERIC)
dropna_ps_tfidf[:3]

# Since process steps are mutually exclusive, we set penalty to -1.
dropna_ps_ohe = return_one_hot(dropna_ps['Process Step'], 0)
dropna_ps_ohe.to_csv('07 NM PS OHE.csv', encoding='utf-8', index=False, quoting=csv.QUOTE_NONNUMERIC)
dropna_ps_ohe[:3]

### NM PT

dropna_pt = pandas.read_csv('06 NM PT.csv', delimiter=',', encoding='latin-1').fillna('')
dropna_pt[:1]

dropna_pt_tfidf = return_tfidf_matrix(dropna_pt['Bony Cleaned'], 1000, 5, (1,3))
dropna_pt_tfidf.to_csv('07 NM PT TFIDF.csv', encoding='utf-8', index=False, quoting=csv.QUOTE_NONNUMERIC)
dropna_pt_tfidf[:3]

# Since process steps are mutually exclusive, we set penalty to -1.
dropna_pt_ohe = return_one_hot(dropna_pt['Problem Type'], 0)
dropna_pt_ohe.to_csv('07 NM PT OHE.csv', encoding='utf-8', index=False, quoting=csv.QUOTE_NONNUMERIC)
dropna_pt_ohe[:3]

### NM CF

dropna_cf = pandas.read_csv('06 NM CF.csv', delimiter=',', encoding='latin-1').fillna('')
dropna_cf[:1]

dropna_cf_tfidf = return_tfidf_matrix(dropna_cf['Bony Cleaned'], 1000, 5, (1,3))
dropna_cf_tfidf.to_csv('07 NM CF TFIDF.csv', encoding='utf-8', index=False, quoting=csv.QUOTE_NONNUMERIC)
dropna_cf_tfidf[:3]

column = dropna_cf['Contributing Factors']
option_set = set([cf for cfs in column for cf in cfs.split('|')])
dropna_cf_ohe = pandas.DataFrame([[1 if option in value else 0 for option in option_set] for value in column], columns = option_set)
dropna_cf_ohe.to_csv('07 NM CF OHE.csv', encoding='utf-8', index=False, quoting=csv.QUOTE_NONNUMERIC)
dropna_cf_ohe[:3]

### NM OS

dropna_os = pandas.read_csv('06 NM OS.csv', delimiter=',', encoding='latin-1').fillna('')
dropna_os[:1]

dropna_os_tfidf = return_tfidf_matrix(dropna_cf['Bony Cleaned'], 1000, 5, (1,3))
dropna_os_tfidf.to_csv('07 NM OS TFIDF.csv', encoding='utf-8', index=False, quoting=csv.QUOTE_NONNUMERIC)
dropna_os_tfidf[:3]

column = dropna_cf['Overall Severity']
option_set = set([cf for cfs in column for cf in cfs.split('|')])
dropna_cf_ohe = pandas.DataFrame([[1 if option in value else 0 for option in option_set] for value in column], columns = option_set)
dropna_cf_ohe.to_csv('07 NM OS OHE.csv', encoding='utf-8', index=False, quoting=csv.QUOTE_NONNUMERIC)
dropna_cf_ohe[:3]
