==> 98 Only Other.ipynb <==
import pandas
import csv

cmb = pandas.read_csv('05 Combined cleaned.csv', delimiter=',', encoding='latin-1').fillna('')
cmb[:5]

pso = cmb[cmb['Process step']=='Other']

pto = cmb[cmb['Problem type']=='Other']

cfo = cmb[cmb['Contributing factors']=='Other']

pso.to_csv('pso.csv', encoding='utf-8', index=False, quoting=csv.QUOTE_NONNUMERIC)

pto.to_csv('pto.csv', encoding='utf-8', index=False, quoting=csv.QUOTE_NONNUMERIC)

cfo.to_csv('cfo.csv', encoding='utf-8', index=False, quoting=csv.QUOTE_NONNUMERIC)
